// artifact/src/lib/types.ts
export type DatabaseType = "mysql" | "postgres";
// ...
export interface Diagram {
  id?: number;
  name: string;
  dbType: DatabaseType;
  data: {
    nodes: AppNode[]; //tables
    edges: AppEdge[]; //relations
    notes?: AppNoteNode[];
    zones?: AppZoneNode[];
    viewport: { x: number; y: number; zoom: number };
    isLocked?: boolean;
  };
  createdAt: Date;
  updatedAt: Date;
  deletedAt?: Date | null;
}
// ...
export interface TableNodeData extends Record<string, unknown> {
  label: string;
  schema?: string;
  qualifiedName?: string;
  columns: Column[];
  indices?: Index[];
  comment?: string;
  color?: string;
  isDeleted?: boolean;
  deletedAt?: Date;
  order?: number;
  isPositionLocked?: boolean;
  onDelete?: (ids: string[]) => void;
  checkConstraints?: CheckConstraint[];
  partitionInfo?: PartitionInfo;
}

export interface NoteNodeData extends Record<string, unknown> {
  text: string;
  color?: string;
  onUpdate?: (id: string, data: Partial<NoteNodeData>) => void;
  onDelete?: (ids: string[]) => void;
  isPositionLocked?: boolean;
}

export interface ZoneNodeData extends Record<string, unknown> {
  name: string;
  color?: string;
  onUpdate?: (id: string, data: Partial<ZoneNodeData>) => void;
  onDelete?: (ids: string[]) => void;
  onCreateTableAtPosition?: (position: { x: number; y: number }) => void;
  onCreateNoteAtPosition?: (position: { x: number; y: number }) => void;
  isLocked?: boolean;
}

export interface EdgeData extends Record<string, unknown> {
  relationship: string;
  isHighlighted?: boolean;
  isPositionLocked?: boolean;
  constraintName?: string;
  sourceColumns?: string[];
  targetColumns?: string[];
  onDelete?: string;
  onUpdate?: string;
  deferrable?: boolean;
  initiallyDeferred?: boolean;
  isComposite?: boolean;
  centerX?: number;
  centerY?: number;
}

export interface Settings {
  rememberLastPosition: boolean;
  snapToGrid: boolean;
  focusTableDuringSelection: boolean;
  focusRelDuringSelection: boolean;
  allowTableOverlapDuringCreation: boolean;
  enableFreePanning: boolean;
  exportForeignKeyConstraint: boolean;
  sidebarScale: number;
  checkpoints: CheckpointSettings;
}

export type AppNode = Node<TableNodeData, "table">;
export type AppNoteNode = Node<NoteNodeData, "note">;
export type AppZoneNode = Node<ZoneNodeData, "zone">;
export type AppEdge = Edge<EdgeData>;